Trim code.
